Added basic React setup for what I want
We now have the side bar mini posts and the main article posts as React
elements. I still need to learn how to utilize a NoSQL type of database
to store articles and dynamically load them onto the website as the user
loads the website onto their browser.

For now, this is good progress because I can now manipulate elements
very specifically from within React.

The next steps are to investigate the database aspect and its
interactions with React.