Fixed the client, added pages
Fixed the problem with the Routes and react-router-dom v6. Also added
all of the needed pages and Routes to them.

I need to add the server aspect to be able to retrieve X number of
articles from the database and display them as React elements. This
information should be on D:/Documents/Workspaces/website/repo