Improving markdown implementation
No longer need multiple markdown files. Now we can simply make a single .mdx file containing BOTH markdown and also HTML and/or React within the .mdx file. This was hard to implement because of configuration issues. In the end it is all worth it because now my pages are all stored within a single file, and I can be as flexible as I need within the markdown, with special features now being a possibility.